+++ src/collective/collectionfilter/baseviews.py
@@ -173,6 +173,7 @@ class BaseSearchView(BaseView):
     def urlquery(self):
         urlquery = {}
         urlquery.update(self.top_request.form)
+
         for it in (
             TEXT_IDX,
             'b_start',